+++ src/hextostring/evaluate/EvaluatorFactory.java
@@ -1,5 +1,6 @@
 package hextostring.evaluate;
 
+import hextostring.evaluate.encoding.EncodingEvaluator;
 import hextostring.evaluate.hex.HexStringEvaluator;
 import hextostring.evaluate.hex.SJISHexStringEvaluator;
 import hextostring.evaluate.hex.UTF16HexStringEvaluator;
@@ -23,8 +24,12 @@ public class EvaluatorFactory {
 		new UTF16HexStringEvaluator();
 	private static HexStringEvaluator utf8HexStringEvaluatorInstance =
 		new UTF8HexStringEvaluator();
+
 	private static ReadableStringEvaluator readableStringEvaluatorInstance =
 		new JapaneseStringEvaluator();
+
+	private static EncodingEvaluator encodingEvaluatorInstance =
+		new EncodingEvaluator();
 
 	/**
 	 * Provides a hex evaluator for a given charset.
@@ -34,7 +39,9 @@ public class EvaluatorFactory {
 	 * @return A fitting evaluator.
 	 */
 	public static HexStringEvaluator getHexStringEvaluatorInstance(Charset cs) {
-		if (cs == Charsets.SHIFT_JIS) {
+		if (cs == Charsets.DETECT) {
+			return null;
+		} else if (cs == Charsets.SHIFT_JIS) {
 			return sjisHexStringEvaluatorInstance;
 		} else if (cs == Charsets.UTF16_BE || cs == Charsets.UTF16_LE) {
 			return utf16HexStringEvaluatorInstance;
@@ -46,6 +53,10 @@ public class EvaluatorFactory {
 
 	public static ReadableStringEvaluator getReadableStringEvaluatorInstance() {
 		return readableStringEvaluatorInstance;
+	}
+
+	public static EncodingEvaluator getEncodingEvaluatorInstance() {
+		return encodingEvaluatorInstance;
 	}
 
 }